Combines component as basic application
Issues with segfaults in meshpy.triangle related to order of of meshpy
vs. numpy imports have emerged again. See this issue on the meshpy repo
(https://github.com/inducer/meshpy/issues/8).

Look into using TetGen to do the 2d triangulation as the issue seems to
be restricted to triangle.
