+++ mesh_sphere_packing/splitsphere.py
@@ -1,66 +1,6 @@
-#!/usr/bin/env python
-from __future__ import print_function
+from mesh_sphere_packing import *
 
-import sys
-
-import numpy as np
-from numpy import random as npr
-from numpy import linalg as npl
 from scipy.spatial.qhull import ConvexHull
-
-
-def plot_segment_points_3d(data):
-    """This function is for visualising sphere points to aid debugging/development
-    """
-    # TODO : remove this after development complete
-    import matplotlib.pyplot as plt
-    from mpl_toolkits.mplot3d import Axes3D
-    fig = plt.figure()
-    ax = fig.add_subplot(111, projection='3d')
-    ax.scatter(data[:,0], data[:,1], zs=data[:,2])
-    plt.show()
-
-
-def plot_polys(polys):
-    """This function is for visualising sphere segments to aid debugging/development
-    """
-    # TODO : remove this after development complete
-    from mpl_toolkits.mplot3d.art3d import Poly3DCollection
-    import matplotlib.pyplot as plt
-    fig = plt.figure()
-    ax = fig.add_subplot(111, projection='3d')
-    fc = ["crimson" if i%2 else "gold" for i in range(len(polys))]
-    ax.add_collection3d(Poly3DCollection(poly3d, facecolors=fc, linewidths=1))
-    plt.show()
-
-
-def write_geomview(points, tris, fname):
-    with open(fname, 'w') as f:
-        f.write('OFF\n')
-        f.write('%d %d %d\n' % (len(points), len(tris), 0))
-        for p in points:
-            f.write('%+1.15e %+1.15e %+1.15e\n' % (p[0], p[1], p[2]))
-        for t in tris:
-            f.write('3 %d %d %d\n' % (t[0], t[1], t[2]))
-
-
-def output_segments_geomview(segments):
-    for i, (tris, points) in enumerate(segments):
-        write_geomview(points, tris, './segment_%d.off' % i)
-
-
-def get_args(argv):
-    """Get command line arguments
-    :return: sphere center coordinates, x, y, z, sphere radius, r,
-    domain box side lengths, Lx, Ly, Lz.
-    """
-    try:
-        return float(argv[1]), float(argv[2]), float(argv[3]), float(argv[4]),\
-                float(argv[5]), float(argv[6]), float(argv[7])
-    except IndexError:
-        raise UserWarning('Must specify x0 y0 z0 r Lx Ly Lz')
-    except ValueError:
-        raise UserWarning('Invalid arguments')
 
 
 def gen_sphere_spiral_points(x, y, z, r, num_points=200):
@@ -222,6 +162,7 @@ def split_sphere_points_segments(points, x, y, z, r, Lx, Ly, Lz):
 
     return segments
 
+
 def translate_segments(segments, Lx, Ly, Lz):
     translations = [
         np.array([0., 0., 0.]),
@@ -233,9 +174,10 @@ def translate_segments(segments, Lx, Ly, Lz):
         np.array([0., Ly, Lz]),
         np.array([Lx, Ly, Lz]),
     ]
-    for (tris, points), t in zip(segments, translations):
+    for (points, tris), t in zip(segments, translations):
         points += t
     return segments
+
 
 def triangulate_segment_points(point_sets, x, y, z):
     """For now this function returns a set of convex hulls
@@ -275,8 +217,8 @@ def triangulate_segment_points(point_sets, x, y, z):
                     points_reindexed[count] = points[idx]
                     count += 1
         return (
+            points_reindexed[:count],
             np.array([[reindex[idx] for idx in t] for t in tris]),
-            points_reindexed[:count]
         )
 
     def extract_surf_from_chull(chull, x, y, z):
@@ -292,34 +234,13 @@ def triangulate_segment_points(point_sets, x, y, z):
     return [extract_surf_from_chull(h, x, y, z) for h in chulls]
 
 
-def triangulate_segment_points_cgal_alt(point_sets):
-    from CGAL.CGAL_Kernel import Point_3
-    from CGAL import CGAL_Convex_hull_3
-    from CGAL.CGAL_Polyhedron_3 import Polyhedron_3
-    """For now this function returns a set of convex hulls
-    :return: list of scipy.spatial.qhull.ConvexHull instances
-    """
-    # TODO : this function needs to strip out only the triangles which
-    #      : correspond to the surface of the sphere from the triangles
-    #      : in the convex hull generated by qhull.
-    res = []
-    for points in point_sets:
-        points_cgal = [Point_3(*x) for x in points]
-        poly = Polyhedron_3()
-        CGAL_Convex_hull_3.convex_hull_3(points_cgal, poly)
-        res.append(poly)
-    return res
-
-
 def smooth_segments(segments):
     # TODO : implement Laplacian smoothing of inner segment vertices
     return segments
 
 
-if __name__ == '__main__':
-    # x, y, z, r, Lx, Ly, Lz = get_args(sys.argv)
-    x, y, z, r, Lx, Ly, Lz = 0., 0., 0., 0.5, 2., 2., 2.
-    #points = gen_sphere_hypercube_points(x, y, z, r)
+def splitsphere(args):
+    x, y, z, r, Lx, Ly, Lz = args
     points = gen_sphere_spiral_points(x, y, z, r)
     points = filter_points(points, r)
     segment_point_sets = split_sphere_points_segments(
@@ -327,5 +248,25 @@ if __name__ == '__main__':
     )
     segments = triangulate_segment_points(segment_point_sets, x, y, z)
     segments = smooth_segments(segments)
-    segments = translate_segments(segments, Lx, Ly, Lz)
+    return translate_segments(segments, Lx, Ly, Lz)
+
+
+def get_args(argv):
+    """Get command line arguments
+    :return: sphere center coordinates, x, y, z, sphere radius, r,
+    domain box side lengths, Lx, Ly, Lz.
+    """
+    try:
+        return float(argv[1]), float(argv[2]), float(argv[3]), float(argv[4]),\
+                float(argv[5]), float(argv[6]), float(argv[7])
+    except IndexError:
+        raise UserWarning('Must specify x0 y0 z0 r Lx Ly Lz')
+    except ValueError:
+        raise UserWarning('Invalid arguments')
+
+
+if __name__ == '__main__':
+    import sys
+
+    segments = splitsphere(get_args(sys.argv))
     output_segments_geomview(segments)
